FIXTURES
Next, we want to fill our database with dummy data for development, called fixtures. We'll want to install the DoctrineFixturesBundle. Edit composer.json as shown in this commit, then run the composer update command like we did for the DoctrineExtensionsBundle.

Also like we did for the DoctrineExtensionsBundle, we need to explicitly ask our application to load this bundle. But notice this time that we load the bundle within an if block. That means we're loading this bundle only in our dev and test environments. In the production environment, this bundle won't be loaded.