FORM CHOICE FIELD
While browsing, you may have noticed an error on pages that would have a form:

"A __toString() method was not found on the objects of type Jobeet\JobBoardBundle\Entity\Category passed to the choice field."

In order to render a menu of categories, the form needs to know how to stringify category objects. The error message informs us of two possible solutions. We can define a __toString method in the Category class, or we can set a form option to define what getter the form should use. For this example, let's set the form option.