LIFECYCLE EVENTS
The latest change doesn't take into account some project requirements, that a user can re-activate a job for an extra 30 days. Fortunately, in our Job entity and database schema, we have an expiresAt field. When a job is created, the expiresAt field can be automatically set to 30 days after the current date.

When we need to do something automatically to an object before Doctrine persists it to the database, we can listen to Doctrine's prePersist lifecycle event.

Once that's done, we can change the index action to use the expiresAt field instead of createdAt to select the active jobs.