PARAM CONVERTER CONT.
When a job expires, even if you know the URL, it must not be possible to access it anymore. Try the URL for the expired job:

jobeet.com.localhost/job/sensio-labs-paris-france-web-developer-expired

Instead of displaying the job, we need to forward the user to a 404 page. But how can we do this as the job is retrieved automatically by the param converter? Previously, we were able to omit the @ParamConverter annotation because we used type hinting. Let's add the @ParamConverter annotation so that we can set the repository_method option. Now, if you try to get an expired job, you will be forwarded to a 404 page.