PARAM CONVERTER
The Symfony Standard Edition comes with the SensioFrameworkExtraBundle, which adds sweet conventions and annotations. We can use the ParamConverter to automatically convert request parameters into objects. The converter tries to get a Job object from the request attributes. (Request attributes come from route placeholders.) If no Job object is found, a 404 response is generated. But if a Job object is found, then a new "job" request attribute is defined. As with any other request attribute, it is automatically injected in the controller when present in the method signature.