SERVICE CONTAINER
In the Job prePersist method, we have hardcoded the number of days for the job to expire. It would have been better to make the 30 days configurable.

This is a spot where Symfony2 diverges significantly from Symfony1. In Symfony1, settings were available through a global sfConfig class. In Symfony2, settings and other dependencies are passed in through arguments. Dependency injection makes our classes less dependent on a global environment. Though, it also makes their construction more complicated; we need to know about and create all the dependencies before creating the objects we need. To solve this problem, Symfony2 provides a dependency injection container. A dependency injection container is an object that knows how to instantiate and configure objects.

Let's change the Job entity to accept an activeDays argument. Then in the dependency injection container, we'll define an active_days parameter and register a job factory service to create and configure job objects. Then any place where we had created a Job object, instead we'll use the job factory service to construct a Job object for us.