THE TEMPLATE CONT.
Right now, the title of all pages is defined in the <title> tag of the layout. But for the job page, we want to provide more useful information, like the company name and the job position. In Symfony, when a zone of the layout depends on the template to be displayed, you need to define a block. Now, in the show template, we can define the content of the title block. For other pages, like the homepage, we just need a generic title. The default title in the layout will be used with any template that does not override the title block.