CATEGORIES ON THE HOMEPAGE CONT.
Let's change the job index action to get the categories, then in the template, we'll iterate through all categories and display the active jobs. For this to work, we need to add the getActiveJobsByCategory() method to the JobRepository class. Notice how we reuse the active jobs query builder.

This is a spot where Symfony2 diverges significantly from Symfony1. In Symfony1, we would execute database queries from an entity (e.g., $category->getActiveJobs()). In Symfony2, entities are meant to be unaware and independent of any persistence mechanism. To execute a query from within a template and in the context of a particular entity, we'll render the output from another controller action.