FORMS CONT.
The form configuration must sometimes be more precise than what can be introspected from the database schema. For example, the email column is a varchar in the schema, but we need this column to be validated as an email. The goal of validation is to tell you whether or not the data of an object is valid. For this to work, we'll configure a list of rules (called constraints) that the object must follow in order to be valid.