FORMS CONT.
The logoPath column will store the filename of the logo associated with the job. We need to validate that the uploaded file is an image in a web format. We'll also need to change the widget to a file input. Though, we won't need to do this last part explicitly. Symfony can "guess" the type of your field from the validation rules.