FORMS
Most of the time, a form has to be serialized to the database. Since Symfony already knows everything about your database model, it can automatically generate forms based on this information. In fact, when we generated CRUD code, form classes were automatically created as well.

The job form is a perfect example to learn to customize forms. By default, the generated form displays fields for all the table columns. But for the job form, some of them must not be editable by the end user. Removing fields from a form is as simple as removing them from the JobType class.