THE EXTEND ACTION
When a job is expiring in less than five days, or if it is already expired, the user can extend the job's active period for another 30 days from the current date. As always, we need to add a new route and create the extend action. Then, update the "Extend" link code in the preview template.