MAINTENANCE TASKS
Even if Symfony is a web framework, it comes with a command line tool. You have already used it to create the default directory structure of the bundle and to generate various files for the model. Adding a new command is quite easy as the tools used by the Symfony command line are packaged in a framework.

When a user creates a job, he must activate it to put it online. But if not, the database will grow with stale jobs. Let's create a task that remove stale jobs from the database. This task will have to be run regularly in a cron job.