FORMS CONT.
When the form is submitted, the logoPath field will be an instance of UploadedFile. It can be used to move the logo file to a permanent location. But we want the logoPath property to store the relative path to the file. To handle the actual file upload in the form, we're going to use a new logoFile field.

We can integrate the file upload into the entity lifecycle (creation, update, and removal) so that the database operation and the moving of the file become atomic. If there is a problem persisting the entity or if the file cannot be moved, then nothing should happen.

The class now does everything we need: It generates a unique filename before persisting, and it moves the file after persisting.